# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Directory of Organizations API",
        "description": "API для работы со справочником организаций, зданий и видов деятельности",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Получить список компаний с пагинацией",
                "operationId": "5f58a2aa7425e8523ee18129e8c2f592",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CompanyWithRelations"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Создать новую компанию",
                "operationId": "27bddcb311ed27eb53b6c88bc21e0cdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Компания создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyWithRelations"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/company/{companyId}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Получить компанию по ID",
                "operationId": "8596dfdbec81400ec447eaf885211a62",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "description": "ID компании",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyWithRelations"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Компания не найдена"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Удалить компанию",
                "operationId": "5c12919f0b993543816005ecb5a05285",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "description": "ID компании",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Компания удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Company deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Компания не найдена"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Companies"
                ],
                "summary": "Обновить данные компании",
                "operationId": "96d9e61f4d0d2f79a38d38aab9f7772e",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "description": "ID компании",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Компания обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyWithRelations"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Компания не найдена"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/company/by-building/{buildingId}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Получить компании по зданию",
                "operationId": "4bef5c7123830da6b5b05229c9cb8129",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyWithRelations"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/company/by-activity/{activityId}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Получить компании по виду деятельности",
                "operationId": "e9db6d0fd7eb298aeee7d12b56202b1f",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyWithRelations"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/company/search": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Поиск компаний по названию",
                "operationId": "186c94b9aa84e955e8d34334531d6f2d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название компании для поиска",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "companies": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CompanyWithRelations"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Found 5 companies"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Activity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "IT-услуги"
                    },
                    "identifier": {
                        "type": "string",
                        "example": "it_services"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
...
            "Building": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Тверская, 10"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    }
                },
                "type": "object"
            },
...
            "CompanyWithRelations": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Ромашка"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "number": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    }
                },
                "type": "object"
            },
            "CompanyStoreRequest": {
                "required": [
                    "name",
                    "building_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "ООО Ромашка"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "activity_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "+79991234567",
                            "+79997654321"
                        ]
                    }
                },
                "type": "object"
            },
            "CompanyUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Обновленное название"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "activity_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            3,
                            4
                        ]
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "+79998887766"
                        ]
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string"
                    },
                    "last": {
                        "type": "string"
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Static API key authentication",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}